:truck: Move textarea to ui subfolder
